backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic
pydantic==2.10.4
python-dotenv==1.0.1
